rename EdgeResolver to operations
